backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
faiss-cpu
python-dotenv
python-docx
fpdf